refactor: PushManager 구조 선택 및 HikariCP 설정 추가
