+++ backend/bottari/src/test/resources/application.yml
@@ -16,6 +16,9 @@ spring:
     username: ${DATABASE_USERNAME}
     password: ${DATABASE_PASSWORD}
     driver-class-name: com.mysql.cj.jdbc.Driver
+    hikari:
+      maximum-pool-size: 2
+      minimum-idle: 0
   sql:
     init:
       mode: always
